+++ CHANGELOG.md
@@ -7,7 +7,21 @@ and this project adheres to [Semantic Versioning].
 
 ## [Unreleased]
 
-## [0.2.0]
+## [0.3.0]
+
+- [PR-103](https://github.com/itk-dev/ai-screening/pull/103)
+  - Add dump-autoload to woodpecker stg
+- [PR-102](https://github.com/itk-dev/ai-screening/pull/102)
+  - Add missing translations to contrib translations
+- [PR-101](https://github.com/itk-dev/ai-screening/pull/101)
+  - Fix bugs in complexity track
+  - Add ordering to project tracks
+  - Change yellow color nuance
+- [PR-100](https://github.com/itk-dev/ai-screening/pull/100)
+  - Add missing translations to contributed modules and themes
+  - Add confirmation step when project state changes in project edit form.
+
+## [0.2.0] - [0.2.1]
 
 - [PR-98](https://github.com/itk-dev/ai-screening/pull/98)
   - Change bullets in webform progress